@@ -1,3 +1,5 @@
 module github.com/golang/glog
 
-go 1.11
+go 1.19
+
+require github.com/google/go-cmp v0.6.0 // indirect
